requirements.txt
datasets==4.8.4
diffusers==0.37.1
lpips==0.1.4
numpy==2.4.4
pandas==3.0.2
Pillow==12.2.0
torch==2.11.0
torchmetrics==1.9.0
tqdm==4.66.5
transformers==5.5.4
accelerate
git+https://github.com/openai/CLIP.git
